get all user points

+++ backend_web/src/models/points_model.rs
@@ -5,8 +5,8 @@ use uuid::Uuid;
 
 #[derive(Debug)]
 pub struct Points {
-    pub id: i8,
+    pub id: i64,
     pub user_id: Uuid,
-    pub points: i8,
+    pub points: i64,
     pub created_at: DateTime<Utc>,
 }
